Add ReStockService

+++ ApiStorePx/DTOs/Product/CreateProductDto.cs
@@ -3,6 +3,6 @@
     public class CreateProductDto
     {
         public string Name { get; set; } = string.Empty;
-        public int SellPrice { get; set; }
+        public decimal SellPrice { get; set; }
     }
 }
