+++ ApiStorePx/Models/Product.cs
@@ -6,9 +6,9 @@ namespace ApiStorePx.Models
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int SellPrice { get; set; }
+        public decimal SellPrice { get; set; }
         public int Stock { get; set; }
-        public int AvgCost { get; set; }
+        public decimal AvgCost { get; set; }
         public DateTime CreateAt { get; set; } = DateTime.Now;
         public DateTime? UpdateAt { get; set; }
         public ICollection<Restock> restocks { get; set; } = new List<Restock>();
