+++ ApiStorePx/Models/Sale.cs
@@ -7,9 +7,9 @@ namespace ApiStorePx.Models
         public int Id { get; set; }
         public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
         public PaymentMethod PaymentMethod { get; set; }
-        public int Total { get; set; }
-        public int Cost { get; set; }
-        public int Profit { get; set; }
+        public decimal Total { get; set; }
+        public decimal Cost { get; set; }
+        public decimal Profit { get; set; }
         public ICollection<SaleItem> Items { get; set; } = new List<SaleItem>();
     }
 }
