+++ ApiStorePx/Models/SaleItem.cs
@@ -7,8 +7,8 @@
         public int ProductId { get; set; }
         public string Name { get; set; } = string.Empty;
         public int Quantity { get; set; }
-        public int SellPrice { get; set; }
-        public int UnitCost { get; set; }
+        public decimal SellPrice { get; set; }
+        public decimal UnitCost { get; set; }
         public Sale? sale { get; set; }
         public Product? product { get; set; }
     }
